requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
jinja2==3.1.6
pydantic==2.11.7
PyYAML==6.0.2
